# src/main/resources/static/v3/api-docs/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"https://cov19api.herokuapp.com","description":"Generated server url"}],"paths":{"/api/v1/update/full":{"get":{"tags":["update-controller"],"operationId":"requestFullUpdate","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelUpdateRequest"}}}}}}},"/api/v1/update/days":{"get":{"tags":["update-controller"],"operationId":"requestUpdateDays","parameters":[{"name":"days","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelUpdateRequest"}}}}}}},"/api/v1/update/date":{"get":{"tags":["update-controller"],"operationId":"requestUpdateDays_1","parameters":[{"name":"date","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelUpdateRequest"}}}}}}},"/api/v1/update/auto":{"get":{"tags":["update-controller"],"operationId":"requestAutoUpdate","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelUpdateRequest"}}}}}}},"/api/v1/locality/stats":{"get":{"tags":["locality-controller"],"operationId":"getStats","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelMapStringInteger"}}}}}}},"/api/v1/locality/regions":{"get":{"tags":["locality-controller"],"operationId":"allRegions","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelRegion"}}}}}}},"/api/v1/locality/regions/search":{"get":{"tags":["locality-controller"],"operationId":"searchRegions","parameters":[{"name":"codes","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"default":[]}},{"name":"names","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"default":[]}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelRegion"}}}}}}},"/api/v1/locality/regions/id/{id}":{"get":{"tags":["locality-controller"],"operationId":"oneRegionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelRegion"}}}}}}},"/api/v1/locality/regions/code/{code}":{"get":{"tags":["locality-controller"],"operationId":"oneRegionByCode","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelRegion"}}}}}}},"/api/v1/locality/departments":{"get":{"tags":["locality-controller"],"operationId":"allDepartments","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelDepartment"}}}}}}},"/api/v1/locality/departments/search":{"get":{"tags":["locality-controller"],"operationId":"searchDepartments","parameters":[{"name":"codes","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"default":[]}},{"name":"names","in":"query","required":false,"schema":{"type":"array","items":{"type":"string"},"default":[]}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelDepartment"}}}}}}},"/api/v1/locality/departments/id/{id}":{"get":{"tags":["locality-controller"],"operationId":"oneDepartmentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelDepartment"}}}}}}},"/api/v1/locality/departments/code/{code}":{"get":{"tags":["locality-controller"],"operationId":"oneDepartmentByCode","parameters":[{"name":"code","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelDepartment"}}}}}}},"/api/v1/hospitalisation/stats":{"get":{"tags":["hospitalisation-controller"],"operationId":"getStats_1","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}}}}}},"/api/v1/hospitalisation/regionalIntensiveCareAdmissions":{"get":{"tags":["hospitalisation-controller"],"operationId":"allRegionalIntensiveCareAdmissions","responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelRegionalIntensiveCareAdmission"}}}}}}},"/api/v1/hospitalisation/regionalIntensiveCareAdmissions/search":{"get":{"tags":["hospitalisation-controller"],"operationId":"searchRegionalIntensiveCareAdmissionByRegionAndDates","parameters":[{"name":"regionCode","in":"query","required":true,"schema":{"type":"string"}},{"name":"noticeDateBegin","in":"query","required":true,"schema":{"type":"string"}},{"name":"noticeDateEnd","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CollectionModelEntityModelRegionalIntensiveCareAdmission"}}}}}}},"/api/v1/hospitalisation/regionalIntensiveCareAdmissions/id/{id}":{"get":{"tags":["hospitalisation-controller"],"operationId":"oneRegionalIntensiveCareAdmissionById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelRegionalIntensiveCareAdmission"}}}}}}},"/api/v1/hospitalisation/intensiveCareAdmissions":{"get":{"tags":["hospitalisation-controller"],"operationId":"intensiveCareAdmissions","parameters":[{"name":"regionCode","in":"query","required":true,"schema":{"type":"string"}},{"name":"days","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"404":{"description":"Not Found","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RegionalIntensiveCareAdmission"}}}}}}}}},"components":{"schemas":{"EntityModelUpdateRequest":{"type":"object","properties":{"accepted":{"type":"boolean"},"acceptedStartTime":{"type":"string","format":"date-time"},"requestBeginDate":{"type":"string","format":"date-time"},"rejectionCause":{"type":"string"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"Link":{"type":"object","properties":{"rel":{"type":"string"},"href":{"type":"string"},"hreflang":{"type":"string"},"media":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"}}},"EntityModelMapStringInteger":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"CollectionModelEntityModelRegion":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}},"content":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelRegion"}}}},"EntityModelRegion":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"regionCode":{"type":"string"},"regionName":{"type":"string"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"Region":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"regionCode":{"type":"string"},"regionName":{"type":"string"}}},"CollectionModelEntityModelDepartment":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}},"content":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelDepartment"}}}},"EntityModelDepartment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"departmentCode":{"type":"string"},"departmentName":{"type":"string"},"region":{"$ref":"#/components/schemas/Region"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"CollectionModelEntityModelRegionalIntensiveCareAdmission":{"type":"object","properties":{"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}},"content":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelRegionalIntensiveCareAdmission"}}}},"EntityModelRegionalIntensiveCareAdmission":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"region":{"$ref":"#/components/schemas/Region"},"noticeDate":{"type":"string","format":"date"},"intensiveCareAdmissionCount":{"type":"integer","format":"int32"},"links":{"type":"array","items":{"$ref":"#/components/schemas/Link"}}}},"RegionalIntensiveCareAdmission":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"region":{"$ref":"#/components/schemas/Region"},"noticeDate":{"type":"string","format":"date"},"intensiveCareAdmissionCount":{"type":"integer","format":"int32"}}}}}}